+++ include/firesteel/scene.hpp
@@ -21,7 +21,7 @@ namespace Firesteel {
 			nlohmann::json scene=nlohmann::json::parse(ifs);
 			ifs.close();
 			if(scene.contains("entities"))
-				for(size_t i=0;i<scene["entities"].size();i++)
+				for(uint i=0;i<scene["entities"].size();i++)
 					entities.push_back(Entity::deserialize(scene["entities"][i]));
 #ifdef FS_PRINT_DEBUG_MSGS
 			LOGF_DBG("Loaded %d entities from \"%s\"",
@@ -32,7 +32,7 @@ namespace Firesteel {
 			nlohmann::json scene;
 
 			nlohmann::json ents=nlohmann::json::array();
-			for(size_t i=0;i<entities.size();i++) ents.push_back(entities[i]->serialize());
+			for(uint i=0;i<entities.size();i++) ents.push_back(entities[i]->serialize());
 			scene["entities"]=ents;
 
 			std::ofstream o(tPath);
@@ -45,9 +45,9 @@ namespace Firesteel {
 		}
 #endif // !FS_NO_JSON
 		void update() {
-			for(size_t i=0;i<entities.size();i++) entities[i]->update();
+			for(uint i=0;i<entities.size();i++) entities[i]->update();
 		}
-		bool removeAt(const size_t& tId) {
+		bool removeAt(const uint& tId) {
 			if(tId>=entities.size()) {
 				LOGF_WARN("Tried to remove entity at %i while max id is %i",tId,entities.size()-1);
 				return false;
@@ -60,7 +60,7 @@ namespace Firesteel {
 #ifdef FS_PRINT_DEBUG_MSGS
 			LOGF_DBG("Removed %d entities", entities.size());
 #endif // FS_PRINT_DEBUG_MSGS
-			for(size_t i=0;i<entities.size();i++) entities[i]->remove();
+			for(uint i=0;i<entities.size();i++) entities[i]->remove();
 			entities.clear();
 		}
 		std::vector<std::shared_ptr<Entity>> entities;
